requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
chardet==3.0.4
idna==2.7
idna-ssl==1.1.0
multidict==4.3.1
pycryptodome==3.9.4
PyYAML==3.13
simplejson==3.16.0
yarl==1.2.6
cachetools
async_lru
